# stacko-mall-interfaces/src/main/java/com/stacko/mall/interfaces/web/dto/StockAdjustRequest.java
package com.stacko.mall.interfaces.web.dto;

public class StockAdjustRequest {
    private int delta;

    public int getDelta() {
        return delta;
    }

    public void setDelta(int delta) {
        this.delta = delta;
    }
}